refactor nodes into separate package
